requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
alembic==1.13.1

# Settings management
pydantic-settings>=2.0.0
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 # PostgreSQL driver

# Testing
pytest>=7.0.0
httpx>=0.24.0 # Required by FastAPI TestClient 